Introduction of Allocationlayer, new Catalog/Snapshot Logic, PolyValues, new UI/UI-Logic and Various Refactorings (#464)

+++ plugins/sql-language/src/main/java/org/polypheny/db/sql/language/ddl/SqlDropTable.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019-2023 The Polypheny Project
+ * Copyright 2019-2024 The Polypheny Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,21 +17,20 @@
 package org.polypheny.db.sql.language.ddl;
 
 
-import static org.polypheny.db.util.Static.RESOURCE;
-
+import java.util.Optional;
 import org.polypheny.db.algebra.constant.Kind;
-import org.polypheny.db.catalog.entity.CatalogTable;
+import org.polypheny.db.catalog.entity.logical.LogicalEntity;
+import org.polypheny.db.catalog.entity.logical.LogicalTable;
+import org.polypheny.db.catalog.exceptions.GenericRuntimeException;
+import org.polypheny.db.catalog.logistic.EntityType;
 import org.polypheny.db.ddl.DdlManager;
-import org.polypheny.db.ddl.exception.DdlOnSourceException;
 import org.polypheny.db.languages.ParserPos;
-import org.polypheny.db.languages.QueryParameters;
 import org.polypheny.db.prepare.Context;
-import org.polypheny.db.runtime.PolyphenyDbContextException;
+import org.polypheny.db.processing.QueryContext.ParsedQueryContext;
 import org.polypheny.db.sql.language.SqlIdentifier;
 import org.polypheny.db.sql.language.SqlOperator;
 import org.polypheny.db.sql.language.SqlSpecialOperator;
 import org.polypheny.db.transaction.Statement;
-import org.polypheny.db.util.CoreUtil;
 
 
 /**
@@ -51,27 +50,30 @@ public class SqlDropTable extends SqlDropObject {
 
 
     @Override
-    public void execute( Context context, Statement statement, QueryParameters parameters ) {
-        final CatalogTable table;
+    public void execute( Context context, Statement statement, ParsedQueryContext parsedQueryContext ) {
+        final Optional<? extends LogicalEntity> entity = searchEntity( context, name );
 
-        try {
-            table = getCatalogTable( context, name );
-        } catch ( PolyphenyDbContextException e ) {
+        if ( entity.isEmpty() ) {
             if ( ifExists ) {
-                // It is ok that there is no database / schema / table with this name because "IF EXISTS" was specified
+                // It is ok that there is no table with this name because "IF EXISTS" was specified
                 return;
             } else {
-                throw e;
+                throw new GenericRuntimeException( "There exists no entity with the name %s and 'IF EXISTS' was not specified", name );
             }
         }
 
-        try {
-            DdlManager.getInstance().dropTable( table, statement );
-        } catch ( DdlOnSourceException e ) {
-            throw CoreUtil.newContextException( name.getPos(), RESOURCE.ddlOnSourceTable() );
+        Optional<LogicalTable> optionalTable = entity.get().unwrap( LogicalTable.class );
+
+        if ( optionalTable.isEmpty() ) {
+            throw new GenericRuntimeException( "The entity with the name %s is not a relational entity % but of type: ", name, entity.get().dataModel );
+        }
+        LogicalTable table = optionalTable.get();
+
+        if ( table.entityType != EntityType.ENTITY && table.entityType != EntityType.SOURCE ) {
+            throw new GenericRuntimeException( "The entity with the name %s is of %s use the appropriate query.", name, table.entityType );
         }
 
+        DdlManager.getInstance().dropTable( table, statement );
     }
 
 }
-
